requirements.txt
django
mysqlclient
djangorestframework
bcrypt
flake8
pytest
pytest-django